Api V1 Database Migrations (#75)

+++ backend/src/services/sampleDataService.ts
@@ -29,7 +29,12 @@ export default class SampleDataService {
     const tenantService = new TenantService(this.options)
     const memberAttributeSettingsService = new MemberAttributeSettingsService(this.options)
 
-    await memberAttributeSettingsService.createPredefined(CrowdMemberAttributes)
+    await memberAttributeSettingsService.createPredefined(
+      MemberAttributeSettingsService.pickAttributes(
+        [MemberAttributeName.SAMPLE],
+        CrowdMemberAttributes,
+      ),
+    )
     await memberAttributeSettingsService.createPredefined(GithubMemberAttributes)
     await memberAttributeSettingsService.createPredefined(DiscordMemberAttributes)
 
@@ -70,28 +75,43 @@ export default class SampleDataService {
 
   /**
    * Deletes sample data
-   * Sample data is defined for all members and activities where attributes.sample = true
+   * Sample data is defined for all members and activities where attributes.sample.crowd = true
    * Sets currentTenant.hasSampleData to false
+   * Also removes settings for attributes.sample.crowd
    */
   async deleteSampleData(): Promise<void> {
-    // deleting sample members should cascade to their activities as well
-    const memberService = new MemberService(this.options)
     const tenantService = new TenantService(this.options)
 
-    const memberIds = await (
-      await memberService.findAndCountAll({
-        advancedFilter: { sample: true },
+    const tenant = await tenantService.findById(this.options.currentTenant.id)
+
+    if (tenant.hasSampleData) {
+      const memberService = new MemberService(this.options)
+      const memberAttributeSettingsService = new MemberAttributeSettingsService(this.options)
+
+      const memberIds = await (
+        await memberService.findAndCountAll({
+          advancedFilter: { sample: true },
+        })
+      ).rows.reduce((acc, item) => {
+        acc.push(item.id)
+        return acc
+      }, [])
+
+      // deleting sample members should cascade to their activities as well
+      await memberService.destroyBulk(memberIds)
+
+      // delete attribute settings for attributes.sample.crowd as well
+      const sampleAttributeSettings = (
+        await memberAttributeSettingsService.findAndCountAll({
+          filter: { name: MemberAttributeName.SAMPLE },
+        })
+      ).rows[0]
+      await memberAttributeSettingsService.destroyAll([sampleAttributeSettings.id])
+
+      await tenantService.update(this.options.currentTenant.id, {
+        hasSampleData: false,
       })
-    ).rows.reduce((acc, item) => {
-      acc.push(item.id)
-      return acc
-    }, [])
-
-    await memberService.destroyBulk(memberIds)
-
-    await tenantService.update(this.options.currentTenant.id, {
-      hasSampleData: false,
-    })
+    }
 
     console.log(`Sample data for tenant ${this.options.currentTenant.id} deleted succesfully.`)
   }
